# backend/app/models.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2025-04-24T10:22:04+00:00

from __future__ import annotations

from datetime import date, datetime
from enum import Enum
from typing import Any, Dict, List, Optional

from pydantic import BaseModel


class StrategySummary(BaseModel):
    id: Optional[str] = None
    name: Optional[str] = None
    latestVersion: Optional[int] = None
    createdAt: Optional[datetime] = None
    updatedAt: Optional[datetime] = None


class StrategyCreateRequest(BaseModel):
    name: str
    description: Optional[str] = None
    tags: Optional[List[str]] = None
    template: Dict[str, Any]


class StrategyDetail(StrategySummary):
    description: Optional[str] = None
    tags: Optional[List[str]] = None


class StrategyVersionSummary(BaseModel):
    id: Optional[str] = None
    version: Optional[int] = None
    createdAt: Optional[datetime] = None
    message: Optional[str] = None


class StrategyVersionCreateRequest(BaseModel):
    message: Optional[str] = None
    template: Dict[str, Any]


class StrategyVersionDetail(StrategyVersionSummary):
    template: Optional[Dict[str, Any]] = None


class Settings(BaseModel):
    initialCash: Optional[float] = None
    commission: Optional[float] = None


class BacktestRequest(BaseModel):
    strategyVersionId: str
    dataSourceId: str
    startDate: date
    endDate: date
    settings: Optional[Settings] = None


class Status(Enum):
    queued = "queued"
    running = "running"


class BacktestResponse(BaseModel):
    backtestId: Optional[str] = None
    status: Optional[Status] = None


class Status1(Enum):
    running = "running"
    success = "success"
    failed = "failed"


class Trade(BaseModel):
    entryDate: Optional[datetime] = None
    exitDate: Optional[datetime] = None
    profit: Optional[float] = None


class Result(BaseModel):
    totalProfit: Optional[float] = None
    trades: Optional[List[Trade]] = None
    dailyPnL: Optional[List[float]] = None


class BacktestResult(BaseModel):
    status: Optional[Status1] = None
    result: Optional[Result] = None